Refactor iqk: Factor out GEMM for i-quants (AVX2/AVX512)
